Show page count depending on total photos
Loading a new page of photos previously caused the page count to show 1, on account of the dependency the field has on `FilteredPhotos`, which is set to null when it's loading.

This commit makes that field calculated based on `TotalPhotoCount`, which is left alone when the request starts, and updated when it returns (which should give the correct value most of the time).